Fixed Bot Start/Stop Error

+++ SteamBot/BotManager.cs
@@ -57,7 +57,7 @@ namespace SteamBot
             for (int i = 0; i < ConfigObject.Bots.Length; i++)
             {
                 Configuration.BotInfo info = ConfigObject.Bots[i];
-                if (info.AutoStart)
+                if (ConfigObject.AutoStartAllBots || info.AutoStart)
                 {
                     mainLog.Info("Launching Bot " + info.DisplayName + "...");
                 }
@@ -247,41 +247,40 @@ namespace SteamBot
 
             public void Stop()
             {
-                if (IsRunning)
+                if (IsRunning && UsingProcesses)
                 {
-                    if (UsingProcesses)
+                    if (!BotProcess.HasExited)
                     {
-                        if (!BotProcess.HasExited)
-                        {
-                            BotProcess.Kill();
-                            IsRunning = false;
-                        }
+                        BotProcess.Kill();
+                        IsRunning = false;
                     }
-                    else
-                    {
-                        if (TheBot != null)
-                        {
-                            TheBot.StopBot();
-                            IsRunning = false;
-                        }
-                    }
+                }
+                else if (TheBot != null && TheBot.IsRunning)
+                {
+                    TheBot.StopBot();
+                    IsRunning = false;
                 }
             }
 
             public void Start()
             {
-                if (!IsRunning)
+                if (UsingProcesses)
                 {
-                    if (UsingProcesses)
+                    if (!IsRunning)
                     {
                         SpawnSteamBotProcess(BotConfigIndex);
                         IsRunning = true;
                     }
-                    else
-                    {
-                        SpawnBotThread(BotConfig);
-                        IsRunning = true;
-                    }
+                }
+                else if (TheBot == null)
+                {
+                    SpawnBotThread(BotConfig);
+                    IsRunning = true;
+                }
+                else if (!TheBot.IsRunning)
+                {
+                    SpawnBotThread(BotConfig);
+                    IsRunning = true;
                 }
             }
